chore: move to eslint v9
