--- packages/eslint-config-payload/eslint-config/configs/jest/index.js
@@ -1,9 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  env: {
-    jest: true,
-  },
-  plugins: ['jest', 'jest-dom'],
-  extends: ['./rules/jest.js', './rules/jest-dom.js'].map(require.resolve),
-  rules: {},
-}
