--- packages/eslint-config-payload/eslint-config/configs/react/index.js
@@ -1,19 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  env: {
-    browser: true,
-  },
-  plugins: ['jsx-a11y', 'react-hooks', 'react'],
-  settings: {
-    react: {
-      version: 'detect',
-    },
-  },
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
-    },
-  },
-  extends: ['./rules/react-a11y.js', './rules/react.js'].map(require.resolve),
-  rules: {},
-}
