--- packages/eslint-config-payload/index.js
@@ -1,5 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-  root: true,
-  extends: ['./eslint-config/index.js'],
-}
