+++ packages/eslint-config-payload/package.json
@@ -11,24 +11,30 @@
   },
   "license": "MIT",
   "author": "Payload <dev@payloadcms.com> (https://payloadcms.com)",
-  "main": "index.js",
+  "type": "module",
+  "main": "index.mjs",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@types/eslint": "8.56.6",
-    "@typescript-eslint/eslint-plugin": "7.3.1",
-    "@typescript-eslint/parser": "7.3.1",
-    "eslint": "8.57.0",
+    "@eslint/js": "9.6.0",
+    "@types/eslint": "8.56.10",
+    "@types/eslint__js": "8.42.3",
+    "eslint": "9.6.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-import-x": "0.5.3",
-    "eslint-plugin-jest": "27.9.0",
-    "eslint-plugin-jest-dom": "5.1.0",
-    "eslint-plugin-jsx-a11y": "6.8.0",
+    "eslint-plugin-jest": "28.6.0",
+    "eslint-plugin-jest-dom": "5.4.0",
+    "eslint-plugin-jsx-a11y": "6.9.0",
     "eslint-plugin-payload": "workspace:*",
-    "eslint-plugin-perfectionist": "2.7.0",
-    "eslint-plugin-react": "7.34.1",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-regexp": "2.3.0"
+    "eslint-plugin-perfectionist": "2.11.0",
+    "eslint-plugin-react": "7.34.3",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-regexp": "2.6.0",
+    "typescript": "5.5.3",
+    "typescript-eslint": "7.15.0"
+  },
+  "devDependencies": {
+    "globals": "15.8.0"
   }
 }
