+++ packages/eslint-plugin-payload/customRules/no-imports-from-exports-dir.js
@@ -1,5 +1,5 @@
 /** @type {import('eslint').Rule.RuleModule} */
-module.exports = {
+export const rule = {
   meta: {
     type: 'problem',
     docs: {
@@ -28,3 +28,5 @@ module.exports = {
     }
   },
 }
+
+export default rule
