+++ packages/eslint-plugin-payload/customRules/no-non-retryable-assertions.js
@@ -1,5 +1,5 @@
 /** @type {import('eslint').Rule.RuleModule} */
-module.exports = {
+export const rule = {
   meta: {
     type: 'problem',
     docs: {
@@ -63,3 +63,5 @@ module.exports = {
     }
   },
 }
+
+export default rule
