+++ packages/eslint-plugin-payload/customRules/no-relative-monorepo-imports.js
@@ -5,7 +5,7 @@
  */
 
 /** @type {import('eslint').Rule.RuleModule} */
-module.exports = {
+export const rule = {
   meta: {
     type: 'problem',
     docs: {
@@ -33,3 +33,5 @@ module.exports = {
     }
   },
 }
+
+export default rule
